Use ReactModal for ModalEditor
