+++ components/nav.jsx
@@ -7,6 +7,7 @@ import { Transition } from '@headlessui/react'
 import { useAuthentication, useLoggedIn, useMyProfile, useContainer, useWebId } from 'swrlit'
 import { useRouter } from 'next/router'
 import Image from 'next/image'
+import ReactModal from 'react-modal';
 
 import { MailIcon } from '../components/icons'
 import NotePicker from '../components/NotePicker'
@@ -42,19 +43,60 @@ function DevTools() {
 
 function CreateButton() {
   const [modalOpen, setModalOpen] = useState(false)
-  const create = (slateJSON) => {
-    console.log('create:', slateJSON)
+  const save = (plateJSOn) => {
+    console.log('save:', plateJSON)
+  }
+  const [createAnother, setCreateAnother] = useState(false)
+  const closeModal = () => setModalOpen(false)
+  const onSubmit = () => {
+    save(value)
+    if (createAnother) {
+      reset()
+    } else {
+      closeModal()
+    }
   }
   return (
     <div className="flex flex-row max-h-9 self-center">
       <button className="flex btn" onClick={() => setModalOpen(true)}>
         Create
       </button>
-      {modalOpen && (
-        <ModalEditor
-          create={create}
-          closeModal={() => setModalOpen(false)}/>
-      )}
+      <ReactModal isOpen={modalOpen}>
+        <div className="fixed w-full h-full top-0 left-0 flex items-center justify-center">
+          <div className="absolute w-full h-full bg-storm opacity-95"></div>
+          <div className="flex-column fixed align-bottom min-w-2/5 min-h-1/5 max-w-4/5 max-h-4/5 overflow-y-auto overflow-x-hidden bg-snow opactiy-100 rounded-lg shadow-xl">
+            <button type="button" className="absolute top-0 right-0 cursor-pointer flex flex-col items-center mt-4 mr-4 text-fog text-sm" onClick={closeModal}>
+              <svg className="fill-current text-mist" xmlns="http://www.w3.org/2000/svg" width="36" height="36" viewBox="0 0 18 18">
+                <path d="M14.53 4.53l-1.06-1.06L9 7.94 4.53 3.47 3.47 4.53 7.94 9l-4.47 4.47 1.06 1.06L9 10.06l4.47 4.47 1.06-1.06L10.06 9z"></path>
+              </svg>
+            </button>
+
+            <div className="flow-root w-full text-left p-4">
+              <ModalEditor savePlateJSON={save} />
+            </div>
+
+            <div className="block flex-none bottom-0 right-0 w-full bg-mist px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse">
+              <button type="button"
+                className="btn"
+                onClick={onSubmit}>
+                Create
+              </button>
+              <button type="button"
+                className="cancel btn"
+                onClick={closeModal}>
+                Cancel
+              </button>
+              <label className="inline-flex items-center">
+                <input className="form-checkbox text-echeveria"
+                  type="checkbox"
+                  checked={createAnother}
+                  onChange={e => setCreateAnother(e.target.checked)} />
+                <span className="ml-2">Create another</span>
+              </label>
+            </div>
+          </div>
+        </div>
+      </ReactModal>
     </div>
   )
 }
@@ -74,7 +116,7 @@ export default function Nav() {
   const [menuOpen, setMenuOpen] = useState(false)
 
   return (
-    <nav className="py-3 px-6 flex flex-col bg-passionflower-dark w-full">
+    <nav className="p-3 pl-6 pr-6 flex flex-col bg-passionflower-dark w-full">
       <ul className="flex justify-between items-center ">
         <li className="flex flex-row">
           <Link href="/">
